// loan-origination/src/app/feature/BNPL/shared/models/data-information.model.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface DataInformation {
  id:                 string;
  tipoIdentificacion: string;
  identificacion:     string;
  primerApellido:     string;
  name:               string;
  fullName:           string;
  celularFormulario:  string;
  emailFormulario:    string;
  fechaExpedicion:    string;
  client:             string;
  ammount:            number;
  term:               number;
  validationStrategy: string;
  fechaSolicitud:     string;
  referenciaPago:     string;
  score:              number;
  feesApprobe:        number;
  rateApprobe:        number;
  approveValue:       number;
  scoreAcierta:       number;
  value:              number;
  causalRechazo:      string;
  respuesta:          string;
  valorCuota:         number;
  amountFinance:      number;
  status:             string;
  fGA:                number;
  platform:           number;
  urlRedirect:        string;
  pagare:             Autorizacion;
  autorizacion:       Autorizacion;
  typeIdentification: TypeIdentification;
}

export interface Autorizacion {
  name: string;
  pag:  number;
  data: string;
}

export interface TypeIdentification {
  id:          string;
  description: string;
}