// loan-origination/src/app/feature/BNPL/shared/models/validation-strategy-response.model.ts
// Generated by https://quicktype.io

export interface ValidationStrategy {
  id:                 string;
  validationStrategy: string;
  questionarie:       Questionarie;
}

export interface Questionarie {
  cuestionario:         string;
  registroCuestionario: string;
  rtaCuestionario:      any[];
  preguntas:            Pregunta[];
}

export interface Pregunta {
  descripcion: string;
  orden:       string;
  respuestas:  Respuesta[];
}

export interface Respuesta {
  descripcion: string;
  orden:       string;
}